入金登録画面　入金登録処理 DBの外部キー制約張り直し

+++ AccountBookMange/EditorViews/Constants/ListDefine.cs
@@ -14,7 +14,7 @@ namespace EditorViews.Constants
         public class DefineItem
         {
             public string content { get; set; }
-            public int? value { get; set; }
+            public long? value { get; set; }
         }
 
         /// <summary>
